pyproject.toml
[project]
name = "metavision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash>=3.0.4",
    "flask-caching>=2.3.1",
    "matplotlib>=3.10.3",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "scikit-image>=0.25.2",
]
